requirements.txt
tensorflow
streamlit
numpy
pillow
pandas
